+++ pysys/baserunner.py
@@ -116,8 +116,9 @@ class BaseRunner(ProcessUser):
 		
 		self.setKeywordArgs(xargs)
 
-		if self.threads == 0:
+		if self.threads <= 0:
 			self.threads = N_CPUS
+		if self.threads > 1: log.info('Running tests with %d threads', self.threads)
 	
 		self.writers = []
 		summarywriters = []